feedback from mafintosh
