+++ lib/stream.js
@@ -126,7 +126,7 @@ module.exports = class UDXStream extends streamx.Duplex {
     this.emit('connect')
   }
 
-  send (buffer) {
+  async send (buffer) {
     if (!this.connected || this._closed) return false
 
     const id = this._allocSend()
